Laget eget VentekriterierRepository.kt istedet for å legge det inn i FagomraadeRepository.kt
