+++ src/main/kotlin/no/nav/sokos/oppdrag/fastedata/repository/FagomraadeRepository.kt
@@ -10,7 +10,6 @@ import no.nav.sokos.oppdrag.config.DatabaseConfig
 import no.nav.sokos.oppdrag.fastedata.domain.Bilagstype
 import no.nav.sokos.oppdrag.fastedata.domain.Fagomraade
 import no.nav.sokos.oppdrag.fastedata.domain.Korrigeringsaarsak
-import no.nav.sokos.oppdrag.fastedata.domain.Ventekriterier
 
 class FagomraadeRepository(
     private val dataSource: HikariDataSource = DatabaseConfig.db2DataSource,
@@ -114,31 +113,6 @@ class FagomraadeRepository(
                     datoFom = row.string("DATO_FOM"),
                     datoTom = row.string("DATO_TOM"),
                     autoFagsystem = row.string("AUTO_FAGSYSTEMID"),
-                )
-            }
-        }
-
-    fun getVentekriterier(kodeFaggruppe: String): List<Ventekriterier> =
-        sessionOf(dataSource).use { session ->
-            session.list(
-                queryOf(
-                    """
-                    SELECT KODE_FAGGRUPPE, TYPE_BILAG, DATO_FOM, BELOP_BRUTTO, BELOP_NETTO, 
-                           ANT_DAGER_ELDRENN, TIDLIGERE_AAR
-                    FROM T_VENT_KRITERIUM
-                    WHERE KODE_FAGGRUPPE = :KODE_FAGGRUPPE
-                    """.trimIndent(),
-                    mapOf("KODE_FAGGRUPPE" to kodeFaggruppe),
-                ),
-            ) { row ->
-                Ventekriterier(
-                    kodeFaggruppe = row.string("KODE_FAGGRUPPE"),
-                    typeBilag = row.string("TYPE_BILAG"),
-                    datoFom = row.string("DATO_FOM"),
-                    belopBrutto = row.string("BELOP_BRUTTO"),
-                    belopNetto = row.string("BELOP_NETTO"),
-                    antDagerEldreEnn = row.intOrNull("ANT_DAGER_ELDRENN"),
-                    tidligereAar = row.boolean("TIDLIGERE_AAR"),
                 )
             }
         }
